refactor some Event code into new files
